moved modal stuf to internal
