+++ libs/designsystem/modal/internal/src/modal-wrapper/compact/index.ts
@@ -1,0 +1,1 @@
+export * from './modal-compact-wrapper.component';
