+++ libs/designsystem/modal/internal/src/modal-wrapper/compact/modal-compact-wrapper.component.scss
