+++ libs/designsystem/modal/internal/src/modal-wrapper/compact/modal-compact-wrapper.component.ts
@@ -2,9 +2,9 @@ import { Component, ElementRef, HostListener, Injector, Input, OnInit } from '@a
 import { firstValueFrom, Subject } from 'rxjs';
 import { WindowRef } from '@kirbydesign/designsystem/types';
 
-import { Modal } from '../../services/modal.interfaces';
 import { ModalConfig } from '../config/modal-config';
 import { COMPONENT_PROPS } from '../config/modal-config.helper';
+import { Modal } from '../../modal.interfaces';
 
 @Component({
   selector: 'kirby-modal-compact-wrapper',
