+++ libs/designsystem/modal/internal/src/modal-wrapper/config/drawer-supplementary-action.ts
