+++ libs/designsystem/modal/internal/src/modal-wrapper/config/index.ts
@@ -1,0 +1,3 @@
+export * from './drawer-supplementary-action';
+export * from './modal-config';
+export * from './modal-config.helper';
