+++ libs/designsystem/modal/internal/src/modal-wrapper/config/modal-config.ts
