+++ libs/designsystem/modal/internal/src/modal-wrapper/index.ts
@@ -1,0 +1,3 @@
+export * from './modal-wrapper.component';
+export * from './config/index';
+export * from './compact/index';
