+++ libs/designsystem/modal/internal/src/modal-wrapper/modal-wrapper.component.scss
