+++ libs/designsystem/modal/internal/src/public_api.ts
@@ -1,2 +1,3 @@
 export * from './modal.interfaces';
 export * from './modal-navigation.service';
+export * from './modal-wrapper/index';
