+++ libs/designsystem/modal/src/modal/index.ts
@@ -1,10 +1,6 @@
-export { Modal } from './services/modal.interfaces';
 export { ModalController } from './services/modal.controller';
-export { ModalConfig } from './modal-wrapper/config/modal-config';
 export { ModalFooterComponent } from './footer/modal-footer.component';
-export { COMPONENT_PROPS } from './modal-wrapper/config/modal-config.helper';
 export { ActionSheetConfig } from './action-sheet/config/action-sheet-config';
 export { ActionSheetComponent } from './action-sheet/action-sheet.component';
 export { ActionSheetItem } from './action-sheet/config/action-sheet-item';
 export * from './alert';
-export * from './modal-wrapper';
