--- libs/designsystem/modal/src/modal/modal-wrapper/index.ts
@@ -1,1 +1,0 @@
-export * from './modal-wrapper.component';
