+++ libs/designsystem/modal/src/modal/services/modal.helper.ts
@@ -3,14 +3,18 @@ import { ModalController } from '@ionic/angular';
 import { WindowRef } from '@kirbydesign/designsystem/types';
 
 import { KirbyAnimation } from '@kirbydesign/designsystem/helpers';
+import {
+  ModalCompactWrapperComponent,
+  ModalConfig,
+  ModalSize,
+  ModalWrapperComponent,
+  Overlay,
+} from '@kirbydesign/designsystem/modal/internal';
 import { AlertConfig } from '../alert/config/alert-config';
-import { ModalCompactWrapperComponent } from '../modal-wrapper/compact/modal-compact-wrapper.component';
-import { ModalConfig, ModalSize } from '../modal-wrapper/config/modal-config';
-import { ModalWrapperComponent } from '../modal-wrapper/modal-wrapper.component';
+
 import { AlertHelper } from './alert.helper';
 
 import { ModalAnimationBuilderService } from './modal-animation-builder.service';
-import { Overlay } from './modal.interfaces';
 
 @Injectable()
 export class ModalHelper {
