+++ libs/designsystem/tsconfig.json
@@ -38,7 +38,8 @@
       "@kirbydesign/designsystem/empty-state": ["empth-state/index.ts"],
       "@kirbydesign/designsystem/dropdown": ["dropdown/index.ts"],
       "@kirbydesign/designsystem/page": ["page/index.ts"],
-      "@kirbydesign/designsystem/modal": ["modal/index.ts"]
+      "@kirbydesign/designsystem/modal": ["modal/index.ts"],
+      "@kirbydesign/designsystem/modal/internal": ["modal/internal/index.ts"],
     },
     "esModuleInterop": true,
     "target": "es2020"
